pyproject.toml
[tool.poetry]
name = "know-net"
version = "0.1.0"
description = ""
authors = ["Arthu, Wian"]
readme = "README.md"
packages = [{include = "know_net"}]

[tool.poetry.dependencies]
python = ">=3.8.1,<3.9.7 || >3.9.7,<4.0"
networkx = "^3.1"
langchain = "^0.0.292"
beautifulsoup4 = "^4.12.2"
openai = "^0.27.8"
faiss-cpu = "^1.7.4"
sentence-transformers = "^2.2.2"
chromadb = "^0.3.29"
torch = "^2.0.1"
loguru = "^0.7.0"
text-generation = "^0.6.0"
lxml = "^4.9.3"
diskcache = "^5.6.1"
tiktoken = "^0.4.0"
streamlit = "^1.24.1"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
